requirements.txt
Pillow==4.2.1

